exo9

+++ kFoldCrossValidation.m
@@ -1,10 +1,10 @@
 function [xSetsTrain, ySetsTrain, xSetsValidation, ySetsValidation] = kFoldCrossValidation(X,Y, k)
 dimension = size(X,2);
 nTrainPoints = size(X,1);
-xSetsTrain = zeros((k-1)/k * nTrainPoints, dimension, k);
-ySetsTrain = zeros((k-1)/k * nTrainPoints, k);
-xSetsValidation = zeros(1/k * nTrainPoints, dimension, k);
-ySetsValidation = zeros(1/k * nTrainPoints, k);
+xSetsTrain = zeros(int64((k-1)/k * nTrainPoints), dimension, k);
+ySetsTrain = zeros(int64((k-1)/k * nTrainPoints), k);
+xSetsValidation = zeros(int64(1/k * nTrainPoints), dimension, k);
+ySetsValidation = zeros(int64(1/k * nTrainPoints), k);
 
 iter = 1;
 for i = 0 : 1/k : (k-1)/k
